auth.py
# auth.py

from fastapi import APIRouter, Depends, HTTPException
from fastapi.security import OAuth2PasswordRequestForm
# from sqlalchemy.orm import Session
# from models import User, Manager  # Make sure to import your User and Manager models
# from database import get_db  # Your database session dependency
# from utils import create_access_token  # Function to create JWT token
import os
from dotenv import load_dotenv
from datetime import datetime, timedelta
from jose import JWTError, jwt

# from pymongo import MongoClient


load_dotenv()

router = APIRouter()

# Fixed username and password for JWT token generation (stored in environment variables)
ADMIN_USERNAME = os.getenv("ADMIN_USERNAME")
ADMIN_PASSWORD = os.getenv("ADMIN_PASSWORD")

SECRET_KEY = os.getenv("JWT_SECRET")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

# Function to create JWT access token
def create_access_token(data: dict, expires_delta: timedelta = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

# Route to login and get JWT token
@router.post("/token")
async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
    if form_data.username != ADMIN_USERNAME or form_data.password != ADMIN_PASSWORD:
        raise HTTPException(status_code=400, detail="Incorrect username or password")

    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    access_token = create_access_token(data={"sub": ADMIN_USERNAME}, expires_delta=access_token_expires)

    return {"access_token": access_token, "token_type": "bearer"}