requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0

tensorflow==2.15.0
keras==2.15.0

scikit-learn==1.4.2
numpy==1.26.4
pillow==10.3.0

nltk==3.8.1

python-multipart==0.0.9
# --- Web Framework ---
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0

# --- Core ML Libraries ---
tensorflow==2.15.0       # For ANN (Keras is part of TensorFlow now)
scikit-learn==1.4.2      # For KNN and classical ML models
torch==2.3.0             # PyTorch for CNNs or advanced models
torchvision==0.18.0      # For pretrained models like ResNet

# --- Image Processing ---
pillow==10.3.0           # Image manipulation (PIL)
opencv-python==4.9.0.80  # Optional but useful for advanced image ops

# --- NLP & Tokenization ---
nltk==3.8.1              # Tokenization & Lemmatization
# You must run: python -m nltk.downloader punkt wordnet stopwords

# --- Utilities ---
numpy==1.26.4
python-multipart==0.0.9  # For file upload in FastAPI
matplotlib==3.8.4        # Optional (for visualization)

# --- Supabase (optional if you use it) ---
supabase==2.0.2
supabase-py==2.3.0       # If using Supabase client SDK (optional)

# --- Logging and Dev Tools ---
loguru==0.7.2
rich==13.7.1             # For beautiful terminal output

# --- Data Handling ---
pandas==2.2.2
annotated-types==0.7.0
anyio==4.5.0
APScheduler==3.10.4
backports.zoneinfo==0.2.1
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
MetaTrader5==5.0.4424
motor==3.6.0
MT5Manager==5.0.4425
numpy==1.24.4
pandas==2.0.3
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
requests==2.32.3
rsa==4.9
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.31.0
httpx==0.28.1
aiofiles==24.1.0
aiohttp==3.10.11
